Moved view.py into app.py for each app, made apps identical structure
