+++ src/yaba/apps/auth/__init__.py
@@ -1,4 +1,3 @@
 '''Authentication app package.'''
 
 from .app import auth_app
-from .views import __import_me__
