+++ src/yaba/apps/auth/app.py
@@ -1,4 +1,6 @@
 from flask import Blueprint
+from flask import render_template
+from flask_login import LoginManager
 
 from .controller import AuthController
 
@@ -12,3 +14,9 @@ auth_app = Blueprint(
     static_url_path='/static/auth',
 )
 controller = AuthController()
+
+
+@auth_app.route('/')
+def login() -> str:
+    '''Get authentication main page.'''
+    return render_template('auth.html')
