+++ src/yaba/apps/budget/__init__.py
@@ -1,4 +1,3 @@
 '''Budget REST API app package.'''
 
 from .app import budget_app
-from .views import __import_me__
