--- src/yaba/apps/budget/coolf.py
@@ -1,11 +1,0 @@
-
-
-def coolf(x: int, y: int) -> int:
-    '''
-    Sum and takeaway one from two digits.
-
-    :param x: first digit
-    :param y: second digit
-    :return: result
-    '''
-    return x + y - 1
