+++ src/yaba/apps/system/__init__.py
@@ -1,4 +1,3 @@
 '''A system app package.'''
 
 from .app import system_app
-from .views import __import_me__
